fix for no files

+++ ganga/GangaDirac/Lib/RTHandlers/RunTimeHandlerUtils.py
@@ -86,10 +86,11 @@ def master_sandbox_prepare(app, appmasterconfig, sharedir_roots=None):
             from GangaCore.GPIDev.Lib.Job import JobError
             raise JobError("InputFiles have been requested but there are objects in the inputSandBox... Aborting Job Prepare!")
         inputsandbox = []
-        (fileNames, tmpDir) = getInputFilesPatterns(job)
+        fileNames, tmpDir = getInputFilesPatterns(job)
         for filepattern in fileNames:
             inputsandbox.append(File(filepattern))
-        shutil.rmtree(tmpDir)
+        if tmpDir:
+            shutil.rmtree(tmpDir)
     if len(inputsandbox) > 100:
         logger.warning('InputSandbox exceeds maximum size (100) supported by the Dirac backend')
         raise GangaException(None, 'InputSandbox exceed maximum size')
